Cell Activity in Lineage Tree (#5)
* Separate static info from dynamic state

* Add tab for activity in cell lineage

* Small cleanup
